requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
frozenlist==1.4.0
gql==3.4.1
graphql-core==3.2.3
idna==3.4
multidict==6.0.4
numpy==1.25.2
openpyxl==3.1.2
pandas==2.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
